+++ score/mw/com/impl/bindings/lola/methods/skeleton_instance_identifier.h
@@ -54,9 +54,9 @@ class hash<score::mw::com::impl::lola::SkeletonInstanceIdentifier>
 
         constexpr auto instance_id_bit_width =
             std::numeric_limits<decltype(skeleton_instance_identifier.instance_id)>::digits;
-        return std::hash<std::uint64_t>{}(
-            (static_cast<std::uint64_t>(skeleton_instance_identifier.service_id) << instance_id_bit_width) |
-            static_cast<std::uint64_t>(skeleton_instance_identifier.instance_id));
+        return std::hash<std::uint64_t>{}((static_cast<std::uint64_t>(skeleton_instance_identifier.service_id)
+                                           << static_cast<std::uint64_t>(instance_id_bit_width)) |
+                                          static_cast<std::uint64_t>(skeleton_instance_identifier.instance_id));
     }
 };
